backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-apscheduler==1.13.1
python-dotenv==1.0.0
pillow==10.2.0
pytesseract==0.3.10
google-generativeai>=0.8.0
pymongo==4.6.1
certifi
opencv-python-headless>=4.8.0
gunicorn==21.2.0
Werkzeug==3.0.1
openpyxl==3.1.2
requests>=2.31.0
cryptography>=41.0.0
reportlab>=4.0.0
flask-jwt-extended>=4.6.0
marshmallow>=3.20.0
email-validator>=2.0.0
